requirements.txt
aniso8601==8.0.0
beautifulsoup4==4.9.3
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cycler==0.10.0
cymem==2.0.3
easydict==1.9
fastai==1.0.61
fastcore==1.0.19
fastprogress==1.0.0
Flask==1.1.2
Flask-Cors==3.0.9
Flask-RESTful==0.3.8
future==0.18.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.2
murmurhash==1.0.2
numexpr==2.7.1
numpy==1.18.5
nvidia-ml-py3==7.352.0
packaging==20.4
pandas==1.1.2
Pillow==7.2.0
plac==1.1.3
preshed==3.0.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
requests==2.22.0
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
soupsieve==2.0.1
spacy==2.3.2
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
urllib3==1.25.10
wasabi==0.8.0
Werkzeug==1.0.1
